_app/requirements_python.txt
selenium
mysql-connector-python
datetime
requests